split into files dealing with one type of image URL each
